Add Easy Java Solution from LeetCode

+++ LeetCode/Java/Easy/367. Valid Perfect Square/Solution.java
@@ -1,9 +1,9 @@
 class Solution {
     public boolean isPerfectSquare(int num) {
-     int start = 1;
-     int end = num;
+     long start = 1;
+     long end = num;
     while(start<=end){
-       int mid = start + ((end-start)/2);
+       long mid = start + ((end-start)/2);
         if (mid * mid == num)
             return true ;
         else if(mid * mid > num)
